(ref: #15) Extracted superclass with methods that won't change
